upgrade-failover: update replica URL behaviour
The replica URL behaviour was recently changed so that when a master node
dies and the hot standby gets promoted, the replica URL would no longer
be reachable and will only point to the new replica once that is up and
running.

Some discussions to restore the old hot standby URL behaviour has been made
but no commitments yet, so until that happens, let's update the documentation
to prevent any confusion. We will update the documentation again once the
old functionality has been restored.