fix/warning: initial capital letter
By running, `make spell` we have spotted a warning.
The warning has not stopped to merge to main.
However, we can spot the warning appearing in future PR's.

With filename as `Unchanged files with check annotations`.

The fix is to take the main branch and make branch out of it.
Fix the spell and merge to main branch again.

The current open PRs should rebase and this warning shall disappear.