Fix understanding about timeout
HAProxy is running in front of Redis and it is
responsible for doing TLS termination.
HAProxy has a timeout of 12 hours and it independent of Redis
timeout. This is not something configurable by the user.

This information was not clear on the previous statement,
so this commit is expanding and explaning that.

Signed-off-by: laysa.uchoa <laysa.uchoa@aiven.io>