Remove `$` added in commands
According to the README of devportal:

Do not include a `$` before a command that the user should run,
because it will get copied into the user's clipboard and cause
the command to fail (this has been a common standard in the past).

For this reason, I'd like to remove from all the commands the symbol.