add/requirements: pygments
Add an updated version of pygments.
This helps to not depend on the sphinix
version or other package to be updated to
use pygments on v2.12.0

Previous versions have issues supporint
terraform highlight.

This avoid this error:

```
WARNING: Could not lex literal_block as "terraform". Highlighting skipped.
```

You can check out more info:
https://github.com/pygments/pygments/issues/2162