cassandra: add help article on tombstones
Add a new help article on Cassandra tombstones, so that we have
somewhere to direct users who receive a user alert about encountering
too many tombstones during queries. Currently, this article is
pretty short; it could definitely go into more detail, but the
overarching advice of "change your queries or change your data model" is
there, which is the most important bit for someone coming there from a
user alert.

Also flesh out the Cassandra section a bit - at the moment there is only
a reference document, so add a section for concepts as well.