[mysql] [platform]: concepts and calculations
To deal with several platform variables and calcuations in multiple locations,
platform-variables was added as an include to several incuments.

This allows a single source for variables that may be subject to change.
Eg.

.. |mysql_max_connections| replace:: :math:`{\times ⌊usable\_memory⌋ } + { extra\_connection }`
.. |mysql_connections_per_<4G| replace:: :math:`{75}`

While allowing multiple documents and locations to remain unaltered if the
values ever change.

The only downside is we loose a bit of blocking and some formatting due to
substitution. Eg. `.. math::` cannot be used. However, it will accept inline:

    :math:`{max\_connections} =` |mysql_connections_per_<4G| |mysql_max_connections|

Several documents have been reworked to accomodate this change.

The document max-number-of-connections, was moved from how-to to concepts.
Users cannot modify this value, so it's an explanation, rather than tuning.

To this we add the mysql tuning document, to help explain the difference between
max_connections and max_concurrency.