Open Issue if linkcheck fails
We are running our ``linkcheck`` on weekly basis
for all the files on devportal to check if we
spot broken links on the documentation.

We can manually check after the job run
if it failed and check why.

You can do this by going to
GitHub Action > Workflows > ``LinkCheck - All files``
and see the error messages.

This could easily go unnoticed. We need an explicit
way to show to the devportal maintainers
if something is wrong with the devportal links.

A proposal is to open a GitHub issue in case
broken links are spotted on the periodical checks.

Signed-off-by: laysa.uchoa <laysa.uchoa@gmail.com>