services-forking: update article to address customer concern
Some customers may want to fork their production clusters. One of
them recently reached out to us asking if there are any additional
load to their production cluster when the fork is being made.

The short answer is no, and this article is now updated to address
this concern from customers by giving a short paragraph to emphasize
that the fork cluster does not communicate with the old cluster at all
and so will not cause any load to it.