Use psf/black to format Python code
We need to format our code following the best practices and
using psf/black is one of the best python formatters out there.

"Black is the uncompromising Python code formatter."

This workflow runs once per week, and open pull requests for your
review. It helps to keep the code clean when importing from code
to docs and from the devportal code itself.

Fixes: https://github.com/aiven/devportal/issues/1226