First use of affix rules for our dictionary
We have legitimate uses of "VPC's" and "VPCs" in our documentation, so
this is the correct time to start using affix rules (or whatever they're
called) in our dictionary.

I propose to copy rules from the Vale en_US dictionary as they are
needed, and only *use* them with caution.

I *did* copy what seems to be relatively standard stuff from the start
of the en_US .aff file - I definitely need to go back and work out if
those actually have an effect in Vale, and if we want them for our text.

You might ask "why are there so few comments in the .aff file" -
technically it's because there is no such thing as a comment in a
Hunspell .aff file, so what look like comments are working by
coincidence, and thus I'm reluctant to do it too much.