Stop Vale complaining about missing ® marks
For some of them, especially some of the Flink complaints, it wasn't
obvious why the error was happening. I've tried to consistently mark up
for each Example, even though it's more than was necessary, so that we
don't have some parts of the document with more ® marks than the other
equivalent parts (specifically, adjacent **Example** blocks).