Enable anonymous session tracking
This changes the Snowplow tracking configuration
by enabling the anonymous session tracking option.
This allows identifying unique sessions of visitors and
sending those session IDs along with the Snowplow
events even with the anonymous tracking enabled.

Upon calling the `clearUserData` function from Snowplow,
we are including the parameters for preserving the user
data and session anonymously.