kafka-connect: fix gcp storage sink perm requirements [HELP-155]
Unlike stated within the current documentation, the role
`Storage Object Creator` is NOT sufficient for the connector
to operate properly. Specifically, the object permission
`storage.objects.delete` is not part of this role although
it is actually required for re-processing.

The change provides more transparency as per the [connector repo]
(https://github.com/aiven/gcs-connector-for-apache-kafka), and
eventually suggests to use a different [IAM role]
(https://cloud.google.com/storage/docs/access-control/iam-roles).

[HELP-155](https://aiven.atlassian.net/browse/HELP-155)