kafka-connect: Prevent bad config of Pg publication as of Debezium 1.9.7
As part of [#inc-504](https://aiven-io.slack.com/archives/C05CY7VUFML),
we found that an upgrade to Aiven for Apache Kafka Connect 3.4
(from Debezium 1.9.5 to 1.9.7) can lead to Pg source connector failing.
This issue can be mitigated or avoided with a better documentation of
how to properly configure the connector.

[HELP-160]