Update Node.js instructions for connecting to PostgreSQL service
Using the full connection URL for connecting
to a PostgreSQL service with Node.js had
the following issue:

"We can notice the “search” param override from the URL,
and then the href being taken from such url into the
“connectionString” for the config object.

This seems to be a workaround/patch that stems from a
node-postgres/pg-connection-string bug:
the connectionString isn’t properly parsed into its ssl components,
an “sslmode” key is retained and probably overrides the
“ssl” block content if present, but then the CA
isn’t properly parsed and causes problem.

What’s the problem to the user? conn.search is overridden,
so if somebody expects additional parameters from
the connection string to be taken into account,
this creates a subtle bug. The code, generally speaking,
is a bit unclear; either you take it as it is, or it can
stop working with all kind of certificate errors.
Also, sslmode=require usually implies
“estabilish TLS but don’t verify certificates”; while the
“config” ssl block will actually try to verify certificates.
The behaviour of the code doesn’t reflect what’s in the connection string."