Kafka datadog: clarify Kafka metrics vs Kafka consumer metrics
Split the chapter of customising metrics into two - first for the main Kafka integration and the other for the  Consumer Integration.

Add an example how to use exclude_topics.

The issue was that documentation failed to mention that the four keys include_topics, exclude_topics, include_consumer_groups and exclude_consumer_groups only affect the consumer integration. And that kafka_custom_metrics applies to Kafka integration.